+++ index.html
@@ -229,6 +229,5 @@
     </div>
   </template>
 <script type=module src="js/main.js"></script>
-<script type =module src="js/similar-offers.js"></script>
 </body>
 </html>
